+++ Changelog.md
@@ -3,4 +3,6 @@
 ### Version 0.1.0
     * first version
 ### Version 0.2.0
-    * Comming Soon
+    * Added Desktop Sizes
+    * Added Zoom In/Out
+    * Fixed Bugs